+++ demo/seminr-interaction.R
@@ -2,19 +2,15 @@
 
 library(seminr)
 
+# First, using the orthogonal method as per Henseler & Chin (2010).
 # Creating our measurement model
 mobi_mm <- constructs(
   composite("Image",        multi_items("IMAG", 1:5)),
   composite("Expectation",  multi_items("CUEX", 1:3)),
   composite("Value",        multi_items("PERV", 1:2)),
-  composite("Satisfaction", multi_items("CUSA", 1:3))
-)
-
-# Interaction constructs must be created after the measurement model is defined.
-# We are using the orthogonalization method as per Henseler & Chin (2010)
-mobi_xm <- interactions(
-  interaction_ortho("Image", "Expectation"),
-  interaction_ortho("Image", "Value")
+  composite("Satisfaction", multi_items("CUSA", 1:3)),
+  interaction_term(iv = "Image", moderator = c("Expectation"), method = orthogonal, weights = mode_A),
+  interaction_term(iv = "Image", moderator = c("Value"), method = orthogonal, weights = mode_A)
 )
 
 # Structural model
@@ -28,7 +24,6 @@ mobi_sm <- relationships(
 # Load data, assemble model, and estimate
 mobi_pls <- estimate_pls(data = mobi,
                          measurement_model = mobi_mm,
-                         interactions = mobi_xm,
                          structural_model = mobi_sm)
 
 summary(mobi_pls)
@@ -45,13 +40,9 @@ mobi_mm <- constructs(
   composite("Image",        multi_items("IMAG", 1:5)),
   composite("Expectation",  multi_items("CUEX", 1:3)),
   composite("Value",        multi_items("PERV", 1:2)),
-  composite("Satisfaction", multi_items("CUSA", 1:3))
-)
-
-# interaction constructs must be created after the measurement model is defined
-mobi_xm <- interactions(
-  interaction_scaled("Image", "Expectation"),
-  interaction_scaled("Image", "Value")
+  composite("Satisfaction", multi_items("CUSA", 1:3)),
+  interaction_term(iv = "Image", moderator = c("Expectation"), method = product_indicator, weights = mode_A),
+  interaction_term(iv = "Image", moderator = c("Value"), method = product_indicator, weights = mode_A)
 )
 
 # structural model: note that name of the interactions construct should be
@@ -65,7 +56,40 @@ mobi_sm <- relationships(
 # Load data, assemble model, and estimate
 mobi_pls <- estimate_pls(data = mobi,
                          measurement_model = mobi_mm,
-                         interactions = mobi_xm,
+                         structural_model = mobi_sm)
+
+summary(mobi_pls)
+
+# Bootstrap the model
+boot_mobi_pls <- bootstrap_model(seminr_model = mobi_pls,
+                                 nboot = 500)
+
+summary(boot_mobi_pls)
+
+library(seminr)
+
+# Third, using the two_stage method as per Henseler & Chin (2010).
+# Creating our measurement model
+mobi_mm <- constructs(
+  composite("Image",        multi_items("IMAG", 1:5)),
+  composite("Expectation",  multi_items("CUEX", 1:3)),
+  composite("Value",        multi_items("PERV", 1:2)),
+  composite("Satisfaction", multi_items("CUSA", 1:3)),
+  interaction_term(iv = "Image", moderator = c("Expectation"), method = two_stage, weights = mode_A),
+  interaction_term(iv = "Image", moderator = c("Value"), method = two_stage, weights = mode_A)
+)
+
+# Structural model
+#  note: interactions should be the names of its main constructs joined by a '*' in between.
+mobi_sm <- relationships(
+  paths(to = "Satisfaction",
+        from = c("Image", "Expectation", "Value",
+                 "Image*Expectation", "Image*Value"))
+)
+
+# Load data, assemble model, and estimate
+mobi_pls <- estimate_pls(data = mobi,
+                         measurement_model = mobi_mm,
                          structural_model = mobi_sm)
 
 summary(mobi_pls)